Tests/LibWeb/Text/expected/HTML/Location-protocol-setter.txt

Setting location.protocol to an invalid value throws SyntaxError
